Reorganize modules with API views
Put api.v1 views into a separate package (core). This will allow to add new
"types" of API endpoints without creating a mess. For example:
- api.v1.core
- api.v1.datasets
- api.v1.evaluation
