+++ webserver/__init__.py
@@ -76,19 +76,27 @@ def create_app_sphinx():
 
 
 def _register_blueprints(app):
-    from webserver.views.index import index_bp
-    from webserver.views.data import data_bp
-    from webserver.views.api_v1 import api_v1_bp
-    from webserver.views.api_legacy import api_legacy_bp
-    from webserver.views.stats import stats_bp
-    from webserver.views.login import login_bp
-    from webserver.views.user import user_bp
-    from webserver.views.datasets import datasets_bp
-    app.register_blueprint(index_bp)
-    app.register_blueprint(data_bp)
-    app.register_blueprint(api_v1_bp, url_prefix='/api/v1')
-    app.register_blueprint(api_legacy_bp)
-    app.register_blueprint(stats_bp)
-    app.register_blueprint(login_bp, url_prefix='/login')
-    app.register_blueprint(user_bp)
-    app.register_blueprint(datasets_bp, url_prefix='/datasets')
+
+    def register_ui(app):
+        from webserver.views.index import index_bp
+        from webserver.views.data import data_bp
+        from webserver.views.stats import stats_bp
+        from webserver.views.login import login_bp
+        from webserver.views.user import user_bp
+        from webserver.views.datasets import datasets_bp
+        app.register_blueprint(index_bp)
+        app.register_blueprint(data_bp)
+        app.register_blueprint(stats_bp)
+        app.register_blueprint(login_bp, url_prefix='/login')
+        app.register_blueprint(user_bp)
+        app.register_blueprint(datasets_bp, url_prefix='/datasets')
+
+    def register_api(app):
+        from webserver.views.api.v1.core import bp_core
+        app.register_blueprint(bp_core, url_prefix='/api/v1')
+
+        from webserver.views.api.legacy import api_legacy_bp
+        app.register_blueprint(api_legacy_bp)
+
+    register_ui(app)
+    register_api(app)
